Extract the configuration_patch to a different file
